requirements.txt
edge-tts>=6.1.9
moviepy>=1.0.3
requests>=2.31.0
Pillow>=10.0.0
flask>=3.0.0
python-dotenv>=1.0.0
